requirements-docker.txt
# Docker-optimized requirements
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database dependencies
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0

# Data processing
pandas==2.1.4
openpyxl==3.1.2

# Rate limiting
slowapi==0.1.9

# Environment & Configuration
python-dotenv==1.0.1

# Templates
Jinja2==3.1.4

# Additional utilities
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Production server
gunicorn==21.2.0

# Security
cryptography==41.0.7
